UI enhancement
Fix error message display

Display storage timestamp

Make collapsable lists scrollable on overflow

Display overlay on loading

Rename Verifier file

Fix typo on landing page
