+++ ui-draft/src/styles/components/_loading-overlay.scss
@@ -2,6 +2,10 @@
   position: absolute;
   width: 100%;
   height: 100%;
+  top: 0;
+  // bottom: 0;
+  left: 0;
+  // right: 0;
   background: $black-500;
   display: flex;
   align-items: center;
